# Speed up appointments JSON responses