--- a/app/assets/javascripts/modules/calendars/company.es6
+++ b/app/assets/javascripts/modules/calendars/company.es6
@@ -26,7 +26,8 @@ class CompanyCalendar extends Calendar {
       eventTextColor: '#fff',
       eventSources: [
         {
-          url: '/appointments'
+          url: '/appointments',
+          cache: true
         },
         {
           url: '/holidays',