--- a/app/assets/javascripts/modules/calendars/my-appointments.es6
+++ b/app/assets/javascripts/modules/calendars/my-appointments.es6
@@ -20,7 +20,8 @@
         eventDataTransform: this.eventDataTransform,
         eventSources: [
           {
-            url: '/appointments?mine'
+            url: '/appointments?mine',
+            cache: true
           },
           {
             url: '/bookable_slots?mine',