--- a/db/migrate/20161208205325_add_start_at_index_on_appointments.rb
+++ b/db/migrate/20161208205325_add_start_at_index_on_appointments.rb
@@ -1,0 +1,5 @@
+class AddStartAtIndexOnAppointments < ActiveRecord::Migration[5.0]
+  def change
+    add_index :appointments, :start_at
+  end
+end