@AndrewVos 2016-12-09
You may need to stop fullcalendar from cache busting. I think by default it sends a _ along.